+++ tests/TestCase/SafeFunctionsTest.php
@@ -1,0 +1,85 @@
+<?php
+/**
+ * This file is part of php-tools.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/php-tools
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       1.0.5
+ */
+namespace Tools\Test;
+
+use PHPUnit\Framework\TestCase;
+use Tools\TestSuite\TestCaseTrait;
+
+/**
+ * SafeFunctionsTest class
+ */
+class SafeFunctionsTest extends TestCase
+{
+    use TestCaseTrait;
+
+    /**
+     * Test for `safe_mkdir()` safe function
+     * @test
+     */
+    public function testSafeMkdir()
+    {
+        $dir = TMP . 'dir_' . md5(time());
+        $this->assertFileNotExists($dir);
+
+        $this->assertTrue(safe_mkdir($dir, 0777, true));
+        $this->assertFileExists($dir);
+        $this->assertTrue(is_dir($dir));
+
+        safe_rmdir($dir);
+    }
+
+    /**
+     * Test for `safe_rmdir()` safe function
+     * @test
+     */
+    public function testSafeRmdir()
+    {
+        $dir = TMP . 'dir_' . md5(time());
+        safe_mkdir($dir, 0777, true);
+
+        $this->assertTrue(safe_rmdir($dir));
+        $this->assertFileNotExists($dir);
+    }
+
+    /**
+     * Test for `safe_symlink()` safe function
+     * @test
+     */
+    public function testSafeSymlink()
+    {
+        $target = tempnam(TMP, 'safe_file');
+        $link = TMP . 'file_' . md5(time());
+        $this->assertFileNotExists($link);
+
+        $this->assertTrue(safe_symlink($target, $link));
+        $this->assertFileExists($link);
+        $this->assertTrue(is_link($link));
+
+        safe_unlink($target);
+        safe_unlink($link);
+    }
+
+    /**
+     * Test for `safe_unlink()` safe function
+     * @test
+     */
+    public function testSafeUnlink()
+    {
+        $file = tempnam(TMP, 'safe_file');
+        $this->assertFileExists($file);
+
+        $this->assertTrue(safe_unlink($file));
+        $this->assertFileNotExists($file);
+    }
+}